09_coding/pyproject.toml
[tool.poetry]
name = "correction-exercices"
version = "0.1.0"
description = ""
authors = ["Nicolas Bancel <nicolas.bancel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
openai = "^1.76.2"
#opencv-python = "^4.11.0.86"
dotenv = "^0.9.9"
pillow = "<10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
